Update boilerplate
- Simplify the app code
- Update all dependencies
- Migrate to webpack 5 & use devServer + localtunnel

+++ plugin-boilerplate/src/app.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
 
-import './styles.less'
+import './styles.css'
 
-function Root() {
+function App() {
   const [state, setState] = React.useState({boardTitle: ''})
 
   async function getBoardTitle() {
@@ -13,12 +13,16 @@ function Root() {
 
   async function deleteAllContent() {
     const allObjects = await miro.board.widgets.get()
+    if (!allObjects.length) {
+      alert('Nothing to delete, try to create some content first')
+      return
+    }
     await miro.board.widgets.deleteById(allObjects.map((object) => object.id))
     await miro.showNotification('Content has been deleted')
   }
 
   return (
-    <div className="container">
+    <div className="container centered">
       <button onClick={() => getBoardTitle()}>Get board title</button>
       <br />
       <div>Board title is: {state.boardTitle}</div>
@@ -29,4 +33,4 @@ function Root() {
   )
 }
 
-ReactDOM.render(<Root />, document.getElementById('react-app'))
+ReactDOM.render(<App />, document.getElementById('root'))
